requirements.txt
Flask==2.0.1
Flask-Bootstrap==3.3.7.1
Flask-Login==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
Pillow==8.4.0
PyMySQL==1.0.2
SQLAlchemy==1.4.23
python-dotenv==0.19.0
waitress==2.0.0
email_validator==1.1.3
requests==2.26.0
contextvars
Werkzeug==2.0.1
Jinja2==3.0.3
WTForms==2.3.3
MarkupSafe==2.0.1
itsdangerous==2.0.1
blinker==1.4
dominate==2.6.0
visitor==0.1.3

# 以下是自动安装的依赖包，不需要手动指定
# click
# colorama
